+++ lib/rggen/core_components.rb
@@ -1,5 +1,18 @@
 require_relative 'core_components/erb_engine'
 
+require_relative 'core_components/code_utility/line'
+require_relative 'core_components/code_utility/code_block'
+require_relative 'core_components/code_utility'
+
+require_relative 'core_components/verilog_utility/identifier'
+require_relative 'core_components/verilog_utility/declaration'
+require_relative 'core_components/verilog_utility/structure_definition'
+require_relative 'core_components/verilog_utility/module_definition'
+require_relative 'core_components/verilog_utility/package_definition'
+require_relative 'core_components/verilog_utility/class_definition'
+require_relative 'core_components/verilog_utility/subroutine_definition'
+require_relative 'core_components/verilog_utility'
+
 require_relative 'core_components/configuration/raise_error'
 require_relative 'core_components/configuration/item'
 require_relative 'core_components/configuration/configuration_factory'
