+++ lib/rggen/core_components/code_utility/line.rb
@@ -1,5 +1,5 @@
 module RgGen
-  module OutputBase
+  module CodeUtility
     class Line
       def initialize
         @words      = []
