+++ spec/rggen/builtins/register/reg_model_spec.rb
@@ -74,7 +74,7 @@ describe 'register/reg_model' do
       end
 
       let(:code) do
-        RgGen::OutputBase::CodeBlock.new
+        RgGen::CodeUtility::CodeBlock.new
       end
 
       let(:expected_code) do
