+++ spec/rggen/core_components/code_utility/code_block_spec.rb
@@ -1,6 +1,6 @@
-require_relative '../../spec_helper'
+require_relative '../../../spec_helper'
 
-module RgGen::OutputBase
+module RgGen::CodeUtility
   describe CodeBlock do
     let(:code_block) do
       CodeBlock.new
