+++ spec/rggen/core_components/verilog_utility/declaration_spec.rb
@@ -1,6 +1,6 @@
 require_relative '../../../spec_helper'
 
-module RgGen::OutputBase::VerilogUtility
+module RgGen::VerilogUtility
   describe Declaration do
     def variable_declaration(attributes)
       Declaration.new(:variable, attributes).to_s
