+++ spec/rggen/core_components/verilog_utility/identifier_spec.rb
@@ -1,6 +1,6 @@
 require_relative '../../../spec_helper'
 
-module RgGen::OutputBase::VerilogUtility
+module RgGen::VerilogUtility
   describe Identifier do
     let(:name) do
       "foo"
