+++ spec/rggen/core_components/verilog_utility/package_definition_spec.rb
@@ -1,6 +1,6 @@
 require_relative '../../../spec_helper'
 
-module RgGen::OutputBase::VerilogUtility
+module RgGen::VerilogUtility
   describe PackageDefinition do
     def package_definition(name, &body)
       PackageDefinition.new(name, &body).to_code.to_s
